Python: configure PYTHONPATH for loading dependencies
We're going to need some python dependencies soon. To make sure they can
be accessed by the runner, setup a directory for installing them and
configure PYTHONPATH so that they will be accessible.

Note that we're not using virtualenv for this purpose because actual
python executable may change (as the user can specify which python
version he wants).